+++ main.c
@@ -21,12 +21,13 @@ int	main(int argc, char **argv)
 {
 	t_map			map_info;
 	t_parse_info	parse_info;
-	
+
 	atexit(hyojocho);
 	validate_args(argc, argv[1]);
-	
+
 	parse_cub_file(argv[1], &map_info, &parse_info);
 	make_array_map(&map_info, &parse_info);
+	engine_main(&map_info);
 	// int i = 0;
 	// while (i < map_info.height)
 	// {
