+++ map/test2.cub
@@ -1,0 +1,20 @@
+NO ./res/north.xpm
+SO ./res/south.xpm
+WE ./res/west.xpm
+EA ./res/east.xpm
+F 220,100,0
+C 225,30,0
+1111111111111111111111111
+1000000000110000000000001
+1000000111000000000000001
+1001000000000000000000001
+1111111110110000011100001
+100000000011000001110111111111111
+ 11110111111111011100000010001
+11110111111111011101010010001
+11000000110101011100000010001
+10000000000000001100000010001
+10000000000000001101010010001
+1100000111010101111101111N0111
+11110111 1110101 101111010001
+11111111 1111111 111111111111
