+++ srcs/engine/hook_utils.c
@@ -1,5 +1,5 @@
 
-#include "cub3d.h"
+#include "../../headers/cub3d.h"
 
 extern int worldMap[14][33];
